pyproject.toml
[project]
name = "strava-data-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "certifi>=2026.2.25",
    "pandas>=2.3.3",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "stravalib>=2.4",
    "streamlit>=1.54.0",
    "truststore>=0.10.4",
]
